feat(sdk): enforce passing an entity for type argument (#13126)
