+++ packages/cli/src/sdk/validate-integration.ts
@@ -3,26 +3,9 @@ import * as errors from '../errors'
 import * as utils from '../utils'
 
 export const validateIntegrationDefinition = (i: sdk.IntegrationDefinition): void => {
-  const { actions, channels, events, states, interfaces } = i
+  const { actions, channels, events, states } = i
 
-  const interfaceActions: Record<string, sdk.ActionDefinition> = interfaces.reduce(
-    (acc, i) => ({
-      ...acc,
-      ...utils.records.mapKeys(i.actions, (_, actionName) => (i.prefix ? `${i.prefix}${actionName}` : actionName)),
-    }),
-    {}
-  )
-
-  const interfaceEvents: Record<string, sdk.EventDefinition> = interfaces.reduce(
-    (acc, i) => ({
-      ...acc,
-      ...utils.records.mapKeys(i.events, (_, eventName) => (i.prefix ? `${i.prefix}${eventName}` : eventName)),
-    }),
-    {}
-  )
-
-  const allActions = { ...(actions ?? {}), ...interfaceActions }
-  const invalidActionNames = _nonCamelCaseKeys(allActions)
+  const invalidActionNames = _nonCamelCaseKeys(actions ?? {})
   if (invalidActionNames.length) {
     throw new errors.BotpressCLIError(
       `The following action names are not in camelCase: ${invalidActionNames.join(', ')}`
@@ -45,8 +28,7 @@ export const validateIntegrationDefinition = (i: sdk.IntegrationDefinition): voi
     )
   }
 
-  const allEvents = { ...(events ?? {}), ...interfaceEvents }
-  const invalidEventNames = _nonCamelCaseKeys(allEvents)
+  const invalidEventNames = _nonCamelCaseKeys(events ?? {})
   if (invalidEventNames.length) {
     throw new errors.BotpressCLIError(`The following event names are not in camelCase: ${invalidEventNames.join(', ')}`)
   }
