+++ packages/sdk/src/integration/index.ts
@@ -1,5 +1,4 @@
 export * from './context'
 export * from './definition'
 export * from './implementation'
-export * from './interface'
 export * from './client'
